requirements.txt
streamlit
pandas
mlflow
pyyaml
scikit-learn
joblib
xgboost
category_encoders
pydantic
seaborn
matplotlib
pytest